Start to migrate envvar tests to integration style
