+++ envvar_test.go
@@ -28,100 +28,6 @@ func TestProjectEnvVarSetInfo(t *testing.T) {
 	}
 }
 
-func TestProjectEnvVarSet(t *testing.T) {
-	server := newFakeServer(t)
-	defer server.stop()
-	appNames := []string{"proj1-dev", "proj1-stage", "proj1-prod"}
-	for _, appName := range appNames {
-		server.prepareResponse(preparedResponse{
-			method:  "POST",
-			path:    "/apps/" + appName + "/env",
-			code:    http.StatusOK,
-			payload: []byte("{}"),
-		})
-	}
-	cleanup, err := setupFakeTarget(server.url())
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cleanup()
-	var c projectEnvVarSet
-	err = c.Flags().Parse(true, []string{
-		"-n", "proj1",
-		"-e", "dev,stage,prod",
-		"--no-restart",
-		"-p",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	var stdout, stderr bytes.Buffer
-	ctx := cmd.Context{
-		Stdout: &stdout,
-		Stderr: &stderr,
-		Args:   []string{"USER_NAME=root", "USER_PASSWORD=r00t", `PREFERRED_TEAM="some nice team"`},
-	}
-	client := cmd.NewClient(http.DefaultClient, &ctx, &cmd.Manager{})
-	err = c.Run(&ctx, client)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expectedOutput := `setting variables in environment "dev"... ok
-setting variables in environment "stage"... ok
-setting variables in environment "prod"... ok
-`
-	if stdout.String() != expectedOutput {
-		t.Errorf("wrong output\nwant:\n%s\ngot:\n%s", expectedOutput, stdout.String())
-	}
-}
-
-func TestProjectEnvVarSetDefaultEnvs(t *testing.T) {
-	server := newFakeServer(t)
-	defer server.stop()
-	appNames := []string{"proj1-dev", "proj1-qa", "proj1-stage", "proj1-prod"}
-	for _, appName := range appNames {
-		server.prepareResponse(preparedResponse{
-			method:  "POST",
-			path:    "/apps/" + appName + "/env",
-			code:    http.StatusOK,
-			payload: []byte("{}"),
-		})
-	}
-	cleanup, err := setupFakeTarget(server.url())
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cleanup()
-	var c projectEnvVarSet
-	err = c.Flags().Parse(true, []string{
-		"-n", "proj1",
-		"--no-restart",
-		"-p",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	var stdout, stderr bytes.Buffer
-	ctx := cmd.Context{
-		Stdout: &stdout,
-		Stderr: &stderr,
-		Args:   []string{"USER_NAME=root", "USER_PASSWORD=r00t", `PREFERRED_TEAM="some nice team"`},
-	}
-	client := cmd.NewClient(http.DefaultClient, &ctx, &cmd.Manager{})
-	err = c.Run(&ctx, client)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expectedOutput := `setting variables in environment "dev"... ok
-setting variables in environment "qa"... ok
-setting variables in environment "stage"... ok
-setting variables in environment "prod"... ok
-`
-	if stdout.String() != expectedOutput {
-		t.Errorf("wrong output\nwant:\n%s\ngot:\n%s", expectedOutput, stdout.String())
-	}
-}
-
 func TestProjectEnvVarSetMissingName(t *testing.T) {
 	var c projectEnvVarSet
 	err := c.Flags().Parse(true, []string{"--no-restart"})
@@ -233,53 +139,6 @@ func TestProjectEnvVarSetErrorInOneOfTheApps(t *testing.T) {
 	}
 	expectedOutput := `setting variables in environment "dev"... ok
 setting variables in environment "stage"... failed
-setting variables in environment "prod"... ok
-`
-	if stdout.String() != expectedOutput {
-		t.Errorf("wrong output\nwant:\n%s\ngot:\n%s", expectedOutput, stdout.String())
-	}
-}
-
-func TestProjectEnvVarSetAppNotFound(t *testing.T) {
-	server := newFakeServer(t)
-	defer server.stop()
-	appNames := []string{"proj1-dev", "proj1-prod"}
-	for _, appName := range appNames {
-		server.prepareResponse(preparedResponse{
-			method:  "POST",
-			path:    "/apps/" + appName + "/env",
-			code:    http.StatusOK,
-			payload: []byte("{}"),
-		})
-	}
-	cleanup, err := setupFakeTarget(server.url())
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer cleanup()
-	var c projectEnvVarSet
-	err = c.Flags().Parse(true, []string{
-		"-n", "proj1",
-		"-e", "dev,stage,prod",
-		"--no-restart",
-		"-p",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	var stdout, stderr bytes.Buffer
-	ctx := cmd.Context{
-		Stdout: &stdout,
-		Stderr: &stderr,
-		Args:   []string{"USER_NAME=root", "USER_PASSWORD=r00t", `PREFERRED_TEAM="some nice team"`},
-	}
-	client := cmd.NewClient(http.DefaultClient, &ctx, &cmd.Manager{})
-	err = c.Run(&ctx, client)
-	if err != nil {
-		t.Fatal(err)
-	}
-	expectedOutput := `setting variables in environment "dev"... ok
-setting variables in environment "stage"... not found
 setting variables in environment "prod"... ok
 `
 	if stdout.String() != expectedOutput {
